IDEX-2334. Move Port Mapping from PropertiesEnvironmentPanel in PropertiesRunnerPanel.
