+++ plugin-runner/che-plugin-runner-ext-runner/src/main/java/org/eclipse/che/ide/ext/runner/client/tabs/properties/panel/impl/PropertiesRunnerPanel.java
@@ -14,12 +14,16 @@ import com.google.gwt.user.client.Timer;
 import com.google.inject.assistedinject.Assisted;
 import com.google.inject.assistedinject.AssistedInject;
 import com.google.inject.name.Named;
+import com.google.web.bindery.event.shared.EventBus;
 
+import org.eclipse.che.api.runner.dto.ApplicationProcessDescriptor;
+import org.eclipse.che.api.runner.dto.PortMapping;
 import org.eclipse.che.ide.api.app.AppContext;
 import org.eclipse.che.ide.api.editor.EditorProvider;
 import org.eclipse.che.ide.api.filetypes.FileTypeRegistry;
 import org.eclipse.che.ide.ext.runner.client.RunnerLocalizationConstant;
 import org.eclipse.che.ide.ext.runner.client.models.Runner;
+import org.eclipse.che.ide.ext.runner.client.runneractions.impl.launch.common.RunnerApplicationStatusEventHandler;
 import org.eclipse.che.ide.ext.runner.client.tabs.container.TabContainer;
 import org.eclipse.che.ide.ext.runner.client.tabs.properties.panel.PropertiesPanelPresenter;
 import org.eclipse.che.ide.ext.runner.client.tabs.properties.panel.PropertiesPanelView;
@@ -31,7 +35,10 @@ import org.eclipse.che.ide.ext.runner.client.util.annotations.LeftPanel;
 
 import javax.annotation.Nonnull;
 
+import java.util.Map;
+
 import static org.eclipse.che.ide.ext.runner.client.constants.TimeInterval.ONE_SEC;
+import static org.eclipse.che.ide.ext.runner.client.runneractions.impl.launch.common.RunnerApplicationStatusEvent.TYPE;
 
 /**
  * @author Andrey Plotnikov
@@ -40,9 +47,10 @@ import static org.eclipse.che.ide.ext.runner.client.constants.TimeInterval.ONE_S
  */
 public class PropertiesRunnerPanel extends PropertiesPanelPresenter {
 
-    private final Timer timer;
+    private final Timer                      timer;
     private final TabContainer               tabContainer;
     private final RunnerLocalizationConstant locale;
+    private final EventBus                   eventBus;
 
     @AssistedInject
     public PropertiesRunnerPanel(final PropertiesPanelView view,
@@ -53,11 +61,13 @@ public class PropertiesRunnerPanel extends PropertiesPanelPresenter {
                                  TimerFactory timerFactory,
                                  @Assisted @Nonnull final Runner runner,
                                  @LeftPanel TabContainer tabContainer,
-                                 RunnerLocalizationConstant locale) {
+                                 RunnerLocalizationConstant locale,
+                                 EventBus eventBus) {
         super(view, appContext);
 
         this.tabContainer = tabContainer;
         this.locale = locale;
+        this.eventBus = eventBus;
 
         // We're waiting for getting application descriptor from server. So we can't show editor without knowing about configuration file.
         timer = timerFactory.newInstance(new TimerFactory.TimerCallBack() {
@@ -90,10 +100,32 @@ public class PropertiesRunnerPanel extends PropertiesPanelPresenter {
 
         this.view.selectShutdown(getTimeout());
         this.view.selectMemory(RAM.detect(runner.getRAM()));
+
+        configureStatusRunEventHandler();
     }
 
     @Override
     public void onConfigLinkClicked() {
         tabContainer.showTab(locale.runnerTabTemplates());
+    }
+
+    private void configureStatusRunEventHandler() {
+        eventBus.addHandler(TYPE, new RunnerApplicationStatusEventHandler() {
+            @Override
+            public void onRunnerStatusChanged(@Nonnull final Runner runner) {
+                setPorts(runner);
+            }
+        });
+    }
+
+    private void setPorts(final Runner runner) {
+        ApplicationProcessDescriptor runnerDescriptor = runner.getDescriptor();
+        if (runnerDescriptor == null) {
+            return;
+        }
+
+        final PortMapping portMapping = runnerDescriptor.getPortMapping();
+        Map<String, String> ports = portMapping != null ? portMapping.getPorts() : null;
+        view.setPorts(ports);
     }
 }
